Shuffle window configuration into its own file.
